Pre-merge reviewer fixes + CodeRabbit findings
Addresses findings from the 👾 internal Reviewer agent pass and a 🐇 CodeRabbit
review round. No src/ hot-path changes; KPI not applicable.

Core:
- Control.h: added comment to min/max fields documenting the three semantic roles (UI clamp, bufSize, natural range) per 👾 finding
- HttpServerModule.cpp: changed %u to %d with explicit (int) cast for int16_t min/max in Uint8 JSON output per 👾 nitpick

Scripts / MoonDeck:
- moondeck.py: narrowed verbatim HTML pass-through to an allowlist of safe structural tags (div, table, span, …) — prevents arbitrary doc content injecting scripts per 👾 finding; added urllib.parse scheme check in _link_tag to strip javascript: and other unsafe schemes per 🐇 finding; moved _allowed_html_re compile outside the per-line loop
- moondeck_ui/app.js: scenario dropdown change handler now calls saveState() to persist selection across refresh per 🐇 finding; postMessage handler restricted to e.source === viewFrame.contentWindow and /api/ URLs only per 🐇 finding; state.scenario documented as intentionally shared across PC + Live cards
- report/history_report.py: removed dead render_commits fallback (unreachable on any working repo) per 👾 finding; empty-rails case now returns a short "no commits" string inline
- run/show_crash_log.py: removed unused sys import per 👾 finding

UI (src/ui/app.js):
- Preview vertex buffer: reuse module-scope vertsBuf (grows but never shrinks) instead of allocating new Float32Array every frame per 👾 finding
- vi === 0 guard: skip lastVerts update on all-black frames so rAF loop idles rather than drawing empty frames per 🐇 finding
- Gamma lift comment: documented why 0.7 (dev-tool readability, not sRGB) per 👾 finding

Docs / CI:
- CLAUDE.md: pre-merge gate 5 triggers Reviewer first for parallel execution; added gate 6 (PR title/description review); renumbered conditional gates 7–9; commit message Reviews category uses 🐇/👾 prefixes
- building.md: aligned "Tested IDF version" to v6.0.0 to match clone command per 🐇 finding

Reviews:
- 👾 Float32Array allocation per frame — fixed (module-scope reuse)
- 👾 Verbatim HTML pass-through too broad — fixed (tag allowlist)
- 👾 ControlDescriptor min/max role overload — fixed (comment added)
- 👾 Dead render_commits fallback — fixed (removed)
- 👾 Unused sys import in show_crash_log.py — fixed (removed)
- 👾 Gamma lift uncommented — fixed (comment added)
- 👾 %u with int16_t — fixed (cast + %d)
- 👾 _serve_markdown_as_html large method — deferred, flag for next subtraction pass
- 👾 atExitHandler design — accepted; premise becomes wrong only if desktop ever calls std::exit() directly, which it doesn't today
- 👾 docs/history/archive/ empty — accepted; plans intentionally removed, decisions.md carries the lessons
- 🐇 building.md version mismatch — fixed (v6.0.0 consistent)
- 🐇 state.scenario not persisted — fixed (saveState() in change handler)
- 🐇 _link_tag unsafe schemes — fixed (urllib.parse scheme check)
- 🐇 vi === 0 empty-frame loop — fixed (early return guard)
- 🐇 postMessage origin — fixed (source + URL validation)

+++ src/core/HttpServerModule.cpp
@@ -360,8 +360,8 @@ void HttpServerModule::writeControls(JsonSink& sink, MoonModule* mod) {
         switch (c.type) {
             case ControlType::Uint8:
                 sink.appendf(
-                    "{\"name\":\"%s\",\"type\":\"uint8\",\"value\":%u,\"min\":%u,\"max\":%u",
-                    c.name, *static_cast<uint8_t*>(c.ptr), c.min, c.max);
+                    "{\"name\":\"%s\",\"type\":\"uint8\",\"value\":%u,\"min\":%d,\"max\":%d",
+                    c.name, *static_cast<uint8_t*>(c.ptr), (int)c.min, (int)c.max);
                 break;
             case ControlType::Uint16:
                 sink.appendf(
